gen_rule/requirements.txt
psycopg2-binary==2.9.9
openai==1.3.0
python-dotenv==1.0.0
